requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
_pytorch_select=0.1=cpu_0
blas=1.0=mkl
ca-certificates=2022.10.11=h06a4308_0
certifi=2022.9.24=py39h06a4308_0
cffi=1.15.0=py39hd667e15_1
flit-core=3.6.0=pyhd3eb1b0_0
intel-openmp=2019.4=243
ld_impl_linux-64=2.38=h1181459_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgomp=11.2.0=h1234567_1
libmklml=2019.0.5=h06a4308_0
libstdcxx-ng=9.1.0=hdf63c60_0
mkl=2020.2=256
mkl-service=2.3.0=py39he8ac12f_0
mkl_fft=1.3.0=py39h54f3939_0
mkl_random=1.0.2=py39h63df603_0
ncurses=6.3=h7f8727e_2
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
numpy=1.19.2=py39h89c1606_0
numpy-base=1.19.2=py39h2ae0177_0
openssl=1.1.1s=h7f8727e_0
pip=22.3.1=py39h06a4308_0
pycparser=2.21=pyhd3eb1b0_0
python=3.9.12=h12debd9_1
pytorch=1.8.1=cpu_py39h60491be_0
readline=8.1.2=h7f8727e_1
setuptools=65.5.0=py39h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.38.5=hc218d9a_0
tk=8.6.12=h1ccaba5_0
typing-extensions=4.4.0=py39h06a4308_0
typing_extensions=4.4.0=py39h06a4308_0
tzdata=2022g=h04d1e81_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h7f8727e_1
zlib=1.2.12=h7f8727e_2
